feat: Добавит базовые ограничения в модуле заданий

+++ project/apps/task-space/src/app/task/task.controller.ts
@@ -4,12 +4,12 @@ import {
   Controller,
   Delete,
   Get,
-  HttpCode,
+  HttpCode, HttpException,
   HttpStatus,
   Param,
   Patch,
   Post,
-  Query,
+  Query, Req,
   UseGuards
 } from '@nestjs/common';
 import { fillObject } from '@project/util/util-core';
@@ -17,7 +17,8 @@ import { TaskRdo } from './rdo/task.rdo';
 import { CreateTaskDto } from './dto/create-task.dto';
 import { UpdateTaskDto } from './dto/update-task.dto';
 import { TaskQuery } from './query/task.query';
-import { JwtAuthGuard } from '@project/shared/authentication';
+import { ExceptionMessages, JwtAuthGuard } from '@project/shared/authentication';
+import { RequestWithPayload, UserRole } from '@project/shared/app-types';
 
 @Controller('tasks')
 export class TaskController {
@@ -32,28 +33,67 @@ export class TaskController {
     return fillObject(TaskRdo, existTask);
   }
 
-  @Get('/')
-  async index(@Query() query: TaskQuery) {
+  @UseGuards(JwtAuthGuard)
+  @Get('/new')
+  async index(@Query() query: TaskQuery, @Req() { user }: RequestWithPayload) {
+    const { sub, role } = user;
+
+    if (role !== UserRole.Executor) {
+      throw new HttpException(ExceptionMessages.FORBIDDEN, HttpStatus.FORBIDDEN);
+    }
+
     const tasks = this.taskService.getTasks(query);
     return fillObject(TaskRdo, tasks);
   }
 
   @UseGuards(JwtAuthGuard)
   @Post('/')
-  async create(@Body() dto: CreateTaskDto) {
-    const createdTask = this.taskService.createTask(dto);
+  async create(@Body() dto: CreateTaskDto, @Req() { user }: RequestWithPayload) {
+    const { sub, role } = user;
+
+    if (role !== UserRole.Customer) {
+      throw new HttpException(ExceptionMessages.FORBIDDEN, HttpStatus.FORBIDDEN);
+    }
+
+    const createdTask = this.taskService.createTask(Object.assign(dto, { userId: sub }));
     return fillObject(TaskRdo, createdTask);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Delete('/:id')
   @HttpCode(HttpStatus.NO_CONTENT)
-  async delete(@Param('id') id: number) {
+  async delete(@Param('id') id: number, @Req() { user }: RequestWithPayload) {
+    const { sub, role } = user;
+
+    if (role !== UserRole.Customer) {
+      throw new HttpException(ExceptionMessages.FORBIDDEN, HttpStatus.FORBIDDEN);
+    }
+
+    const existTask = await this.taskService.getTask(id);
+
+    if (sub !== existTask.userId) {
+      throw new HttpException(ExceptionMessages.FORBIDDEN_UPDATE, HttpStatus.FORBIDDEN);
+    }
+
     await this.taskService.deleteTask(id);
   }
 
+  @UseGuards(JwtAuthGuard)
   @Patch('/:id')
-  async update(@Param('id') id: number, @Body() dto: UpdateTaskDto) {
-    const updatedTask = await this.taskService.updateTask(id, dto);
+  async update(@Param('id') id: number, @Body() dto: UpdateTaskDto,  @Req() { user }: RequestWithPayload) {
+    const { sub, role } = user;
+
+    if (role !== UserRole.Customer) {
+      throw new HttpException(ExceptionMessages.FORBIDDEN, HttpStatus.FORBIDDEN);
+    }
+
+    const existTask = await this.taskService.getTask(id);
+
+    if (sub !== existTask.userId) {
+      throw new HttpException(ExceptionMessages.FORBIDDEN_UPDATE, HttpStatus.FORBIDDEN);
+    }
+
+    const updatedTask = await this.taskService.updateTask(id, Object.assign(dto, { userId: sub }));
     return fillObject(TaskRdo, updatedTask);
   }
 }
